+++ NEWS.md
@@ -1,4 +1,14 @@
-# glsnls 1.2.0
+# gslnls 1.3.0
+
+* Missing starting values/ranges allowed in `gsl_nls()`
+* `lower` and `upper` parameter constraints included in `gsl_nls()` 
+* Added 3 regression problems from Bates & Watts (1988)
+* Updated multi-start algorithm in `gsl_nls()`
+* Added configure.win, cleanup.win and Makevars.win.in
+* Removed old Makevars and Makevars.win
+* Several minor changes
+
+# gslnls 1.2.0
 
 * Added multi-start algorithm to `gsl_nls()`
 * Added 56 NLS regression and optimization test problems